add successful submission state for contact form and cleanup files
